Refactored Interfaces into dto , service , queue
