+++ node-service/services/scan-scheduler.service.ts
@@ -6,11 +6,6 @@ import OnboardingConfig from '../models/OnboardingConfig';
 import Repository from '../models/Repository';
 import logger from '../utils/logger';
 import scanProcessManager from '../services/scan-process-manager.service';
-import rabbitMQService, { QUEUE_SCAN_REPOSITORY } from './rabbitmq.service';
-import {
-  ScanRepositoryMessage,
-  ScheduledScanMessage,
-} from '../types/queue-messages.types';
 
 /**
  * Class for scheduling and managing automatic dependency scans based on user configurations
