+++ node-service/services/user.service.ts
@@ -2,14 +2,7 @@
 import User from '../models/User';
 import { IUser } from '../types/models';
 import { Types } from 'mongoose';
-
-interface GithubUserData {
-  id: number;
-  login: string;
-  name?: string;
-  email?: string;
-  avatar_url?: string;
-}
+import { GithubUserData } from '../types/dto';
 
 export const findOrCreateGithubUser = async (
   githubUserData: GithubUserData,
