+++ node-service/services/workers/vulnerability-scan.worker.ts
@@ -3,7 +3,7 @@ import logger from '../../utils/logger';
 import rabbitMQService, {
   QUEUE_VULNERABILITY_SCAN,
 } from '../../services/rabbitmq.service';
-import { VulnerabilityScanMessage } from '../../types/queue-messages.types';
+import { VulnerabilityScanMessage } from '../../types/queue';
 import vulnerabilityScanService from '../../services/vulnerability-scan.service';
 import Scan from '../../models/Scan';
 import scanProcessManager from '../../services/scan-process-manager.service';
